Fix default size to int

+++ hv/ui/main_window.py
@@ -50,9 +50,9 @@ class HVWindow(QMainWindow):
     def init_size(self):
         desktop = QDesktopWidget()
         size = desktop.availableGeometry().size()
-        width = min(size.width()*0.8, 1280)
-        height = min(size.height()*0.8, 720)
-        size =QSize(width, height)
+        width = min(int(size.width()*0.8), 1280)
+        height = min(int(size.height()*0.8), 720)
+        size = QSize(width, height)
         size = self.settings.value("window_size", size)
         self.resize(size)
